Separated classes in the web package into the subpackages of the dto / controllers
