+++ src/main/java/ru/nsu/fit/service/LoadService.java
@@ -7,9 +7,7 @@ import ru.nsu.fit.database.entities.FileData;
 import ru.nsu.fit.database.entities.Folder;
 import ru.nsu.fit.database.entities.User;
 import ru.nsu.fit.database.repositories.FileDataRepository;
-import ru.nsu.fit.web.login.UserDTO;
-import ru.nsu.fit.web.navigation.FileDTO;
-import ru.nsu.fit.web.navigation.FolderDTO;
+import ru.nsu.fit.web.dtos.FileDTO;
 
 import java.util.Arrays;
 import java.util.List;
@@ -36,7 +34,7 @@ public class LoadService {
         return file;
     }
 
-    public File uploadFileData(int fileId, byte[] data){
+    public File uploadFileData(int fileId, byte[] data) {
         File file = navigationService.getFileInfo(fileId);
         fileDataRepository.save(new FileData(file, data));
         return file;
