+++ src/main/java/ru/nsu/fit/web/dtos/UserPublicDTO.java
@@ -1,4 +1,4 @@
-package ru.nsu.fit.web.navigation;
+package ru.nsu.fit.web.dtos;
 
 import ru.nsu.fit.database.entities.User;
 import ru.nsu.fit.database.types.Role;
